Frontend app sync up the API endpoints
