+++ cccm-frontend/src/components/DashboardPOView.vue
@@ -63,7 +63,7 @@
           <!--Customize the Name field, making it clickable-->
           <template v-slot:item.clientName="{ item }">
             <div class="w-100 h-100">
-              <a :href="`${baseURL}clientrecord/${item.clientID}/${item.csNumber}`">{{item.clientName}}</a>
+              <a :href="`${baseURL}clientrecord/${item.csNumber}`" @click="selectClient(item.csNumber)">{{item.clientName}}</a>
             </div>
           </template>
           <!--Customize the alerts field, show the alert count -->
@@ -121,10 +121,9 @@
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
-import {clientSearch, photoSearch} from "@/components/form.api";
+import {dashboardPOSearch, photoSearch} from "@/components/form.api";
 import { Form } from 'vue-formio';
 import templateClientProfile from '@/components/common/templateClientProfilePO.json';
-
 
 export default {
   name: 'RNAList',
@@ -171,22 +170,29 @@ export default {
   },
   mounted(){
     //form search from the backend
-    this.clientSearchAPI(this.$route.params.poID)
+    this.dashboardPOSearch(this.$route.params.poID)
   },
   methods: {
+    selectClient(csNumber) {
+      console.log("view client [csNumber]: ", csNumber);
+      this.$router.push({
+        name: '${baseURL}clientrecord',
+        params: {csNumber: csNumber}
+      });
+    },
     expandRow ({ item, value }) {
       // call searchPhotoAPI only when the photo hasn't loaded.
       if (this.initDataArray != null && this.initDataArray[item.clientID] != null 
         && this.initDataArray[item.clientID].data != null && this.initDataArray[item.clientID].data.photo != null) {
         return;
       }
-      this.handlePhotoSearch(item.clientID);
+      this.handlePhotoSearch(item.csNumber);
     },
     print() {
       console.log("Print client list");
     },
-    async clientSearchAPI(POID) {
-      const [error, response] = await clientSearch(null, null, null, null,
+    async dashboardPOSearch(POID) {
+      const [error, response] = await dashboardPOSearch(null, null, null, null,
             null, null, null, POID, null);
       //this.initData = response.data;
       this.key_results++;
@@ -402,9 +408,9 @@ export default {
         console.error(error);
       }       
     },
-    async handlePhotoSearch(clientID) {
-      //console.log("Photo search for clientID: ", clientID);
-      const [error, response] = await photoSearch(clientID);
+    async handlePhotoSearch(csNumber) {
+      //console.log("Photo search for csNumber: ", csNumber);
+      const [error, response] = await photoSearch(csNumber);
 
       // To be deleted, sample data
       let photoArary = ["/9j/4AAQSkZJRgABAQEASABIAAD/2wBDABALDA4MChAODQ4SERATGCgaGBYWGDEjJR0oOjM9PDkzODdASFxOQERXRTc4UG1RV19iZ2hnPk1xeXBkeFxlZ2P/xADSAAABBQEBAQEBAQAAAAAAAAAAAQIDBAUGBwgJCgsQAAIBAwMCBAMFBQQEAAABfQECAwAEEQUSITFBBhNRYQcicRQygZGhCCNCscEVUtHwJDNicoIJChYXGBkaJSYnKCkqNDU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6g4SFhoeIiYqSk5SVlpeYmZqio6Slpqeoqaqys7S1tre4ubrCw8TFxsfIycrS09TV1tfY2drh4uPk5ebn6Onq8fLz9PX29/j5+v/AAAsIAyAC1wEAEQD/2gAIAQAAAD8A7+iiiiiiiiiiiiiiiiiiiiiikopKT7o4phdQvz8VyXiHT7O4UsZ1zXA3VoInIj+YVV8tx/DS/MGHHNdBo+qXFqB5cOa6q08TSKB50eytuLU7W8gwJFzjpXnniuydLoOqkr60mjaxNp+AsZ2jvXoGkatFqEC8gN6VrLwMUAAUACl9qMUdqBS0UtFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFJRS0lNyKzdR1OGxTJIyK5W61q8v22Wy7l9qLLQ7q9b/SsqK2I/B9ko5bNP/wCERs/wqJvBdkenWr1l4dtbUcAGm6h4eguk44rmLrR7nSn3QZIFZWpa08w8uZQCOK2PDFpZ3lr5chGTUWo29x4fvfNgz5Yrr9D1RL+0Rifmx0rUqvJdxQuELc1OCpGVNOBopRRRRS0UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUlFFNzWfquoRWVqxc4OOK85bUZNW1EQtkp0rvNF0m2tLdSF5rZ4UcAUbQ1KMDilwBSCk6dKjmSORdsgFcrr3hWGZC8C81gaDZ3djqiIwwld5qWnpqVlsIzxXH6bK2laoYTwg4Fd7G/mWocdxXCeIr6aPVoUQ8V2ekOz2a7vSr3ajtSiiiiloooooooooooooooooooopKKKKWiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiiikooo6CgdKSmE4HsK4Lxpcm4nWGE9eKTw14aZcXDEAiux+1QWcWHYZFVm8Q2iDqKoy+LLVThaavi+1xS/wDCXW3egeLrbPtUy+K7Q4FaVrqNveAbWFXThhjjFQ/Ybffu2DdU/wBwYFct4p0weWJoVw3UkVLoOosbMxyn7q965XU5RNrUefWvRNMG21UD0q50HNJSilozRRRRmijNFFFGaTNLSZpRRRRmjNGaKKKTNFFFLmiiiloopKKWiiiiiiiiiiiiiiiiiiiiiiiiiikoopMZGKb/ALNOxxWXrOox2Nq2SMkV5a9/JNqPmY3AGtmbxDcrGEhidRjtWJdavdS8FiKp/aZz/wAtKZvf+9Sbm/vUm9vWje/96n+Y+OpFXLDVrizb77EV2GkeMFk2pL+tdfZ30Nyg2MM1a46VHcwLNCVYV5trcz6ZOwTIBrBtrtnv43avWtEl86xBHpWivK807FJijOKTrSgYpaBSUtFJmjNGKMYpaM0UUlGaB9KWkozRSYpaKM0tJRijFKKWiikopaKKKKKKKKKKKKKKKKKKKKKKKKKKSikzSVG0iRcsQBVRtVs0yPPXIFcF4ivJNRuRFbnI6cVq6B4TQRB5a6RdDtRD5exfriuV1/wkEBli/IVw80TQSbGGMUzFNzSUo4pKUHHalB2/d4rZ0fXprCRedwFeiaPr0OpRruIU+lbnBTiuK8aaYHj3gVxFlAPM/wB2vUPCbg6fjNbuKKMUmfalGKSnUmKWiiim/WjnsKMhRzxUL3tun3pKrHWLRePMWmHXLQD/AFi1Ul8S26fddag/4SlDwMVIPEqe1IfEqD0oTxNF6irKeILc/wASip01qzPWRR+NTrqdm3/LZakF1bn/AJaipBIh6U8GjFAFLRSYpQKWiiikpaKKKKKKKKKKKKKKKKKKKKKKKKKKSikqKaRYYtzcAV594k8Rv5hihbj2rlPtcjnLO2av6DdpBdgydK9XsJ4bm3UxEHAq0cU10WSPaw4rzjxho/ksZY14rjx6UmKQUo9qSlzRTsDb71asL+WxkDoTx2r0Hwz4lF3iOZgK2Ndh+0WfHPFcCmmNDHKdtdH4NutsXlGuwz6UvasvUNZtrJeWAIrkdR8azI2LfBFR2HjWdmAmwBXY6Vq8d6gwQa1PpS5pe1IDS0x3WJfmOBWfPrNnEPmkrGuvE5H/AB7/ADVly6xrNxwkJxRDZ6nc/wCuRhV+PwwXA3MwqdfB8J6ufzpw8G2v99qePCVsvQ1Kvha37tSHwnan+I1E3hSyTneaqz6JpsK/NNg1zOpx2sH+qmJ/Gsb7dLGfkkNTx69dJ3NX4fGF7HirqeOroDGBVu28auxG/AFbVp4ts2A8yQVoQ+IbCXhZRV+K7hlHyMKsDmijNGaM0maWilooooooooooooooooooooooooopKKSuX8V6mLe2aJTjIrzCRzLJuNMYAUJ/Kuj8O+IJrKdIy3yGvTrK6ivIVePnirBIxWL4mt1lsOnavJZ12XLj0NRYzScYoHtRiikoFLU9tPJauGjOK9B8PeIY7mEQ3R7Yraexglt2KAYNc/ZqtlqqxrwtdvG25QR0xWLrmtw6fERuw1eYalqM17OzFvl7VnnHakrd8P6pJa3Kpn5a9V0+4E1oje1WG6DFKOlNLBF+fGKw9U8QwWg2qea55tU1DUW2wZ21ZtPD89zzcg1t2vhuzhAJWtSGzhiXCKKlACdqcMdqWkyo6kVWmv4IB8zCsi78U2cPAbmuevfGjdIGrFuPFWoOfvcVQn1m5n++1UnkaX72TTAKmS2kk+6tWo9Hun6LV2DwxeSfw1oQeDLk/eWrsfgp+6mph4VlgX92DUDwaxaH930FT2+salBxPWpbeKYFwJjW1aapbXYHlN1q6DRRij2paKKKKWikopaKKKKKKKKKKKKKKKKKSimn7leX+NpnF9szxXLfSgUcDpS9OR2rq/CuvtbusLk4r0a3uYrmMFSKo68CLLg9q8iu+Lt6S3tpLhsIK0W8PXKw7yKypI2gfaaZmjtQBxTkGRwKNjDqKP5VJbzvBIpXjFek+FtajuLby3b2qrqcW3WEdPu1vy3q2mmb8jpXmGuai97csM5FZfKikpwXIp8DbZBXqnhaXzLRFb0roTnscVQ1PU4dPhyxGcVyE+t3GotstiR2q7pvhuSfEl3yK6W10q1tlAjjAxVwAKMCl4owMcU3HrSM6RrngVj3/iK2s15xXKah4x83iDIrnLnWbyc/6w4qi8sjn5zTMUvXjFTQ2cs+Ai1uad4WuZ8EiuhtPBgXHmKK3LTw5aQgZjFXk060TpGKnSGJPuripRgUUtRNFE/Vaqy6XaydUFYeo+FUlOYVArCm07UtLP7tjgelS2PiWe2cLcbjXWWGu292q4ODWoGDAbSKXpzQKUUtFFFFLSUtFFFFFFFFFFFFFFFFFJRTHGI68r8bEf2jiubXjoKcY3x0NMHHagCpYyYx8vDVuaR4jltMKxJArdvvE6TWW3rxXM2WlTand7guFJruNL0G30+MGXb+NbKfY54/KXbXBeLNF+zzGZVwtclRTl9K7Lwtoa3UIZlrW1Twgrw5jwD7VyN74fntP4WwPascxsrYYYqxaXktm4MZwB6VrNr3mYP8QqC712SaDyu1Y+M80h7UvAGAKmhs5pB8qHH0pktvJAfmWvSfBcqm2UZ6Cr+t64lgu1OW9q5hIrrXbgH5lT3rrNK0GGyQZUFq2NuBgcUYwKQHPFO7U3Ixk8VlajrUFgmdwOPeuI1bxZJcErF8ormZrmaY/MxNQ5HpQKMY4q/Z6XPd42qfyrp9K8HF8eZXV6b4ftrNPmRTWokcUXCIBT6cKMCjgdqB9KWkzSClxTNpB68UjwRSj5lDVk33h62uB8iKtcrqPh28sm8yB2AHYU7T/Es9i6xXCk44rs9O1OC9iBUge1Xs+lKMUtFFFFFLRRRRRRRRRRRRRRRRRRSUlJJ/qzXlHjNW/tPGKv8Ahvw2LuNXccV0j+Ebfy8DFcnrfheS1+aNTiuYZSjbWGCKMbaTr7VpaHZNf3Qj5wK9OsbCPTbMbUGcVk3cVzqD4UFR7U7TNKntbpcsxFX/ABXB52klQuTXk00Zik2njFNFWtLtTcXSKBkV61oliLOzGB2q9MxWHdjPtXK3muKZTBNAqr0zisPV9DE0fnQD8q5V0aJtjCmqMc0h5NJ1p8cZYhVFdPonhWS6KtKu0V3VjoNvbRBSoNYnijw+htt8S8+1czpmrSaW3kge1ben6ZLq0wlmyF967Szs4raIKiirHTilpMUduBVe5uorSPLuBXFa74r6pCfyrjbq/muW+ZjVOnZwKRaniikl+SNc/Suo0Pwo0+15Riu60/SoLGIKEBx7VoKoHQYp2KMYFFAGKM4oBpaTFLRTcUuMjFNC7e9GykZEYbWANYOr+GobtMoAp9q5GSK90Gf92HdBXT6D4jW9Cxz4Q10qFcfJ0p2aXNLRRRS0UUUUUUUUUUUUUUUUUUlFI3SvPPGFp/pwlxwK6XwkynTF24rbO4dKjkhjnj2Morg/E/hgQZlgXjrXFOrdx0pld94G09QwlI6iu5eMEfSsTVdYg00Yjxmk0TWf7RHzAVt3EIlh2EcYryXxPZeRfOQMVh9BXX+CrAyyCQivSlAVAvtQQCu2sHWtBiu49yffFYWl3UltMbW6XCdBUfiTw8HUTQLwR2ripYmibYRzTdm1afbW0lw4WNc13HhvwqPlknXGK7eKNYkCIowOKkx8vHWql+R9l5rjbXQEu9TLkcV21rbLbQrGoHAxU6qEHFL/ADpP50dsVmarq8VhF1G6vOtd8RTXz7VPHtXP5J+9SH2o7UuRV3T9OlvHCopxXoGgeF47aNXkX5veuqiiSJcAU7FANOoopOaOKBgCgUcUnTpSiiilpu7FHBo4/Cql1ZRXa7Sowe9cdrOgvYN59oMkelO0HxJskEF2duOOa7aGWOaMOh4qXg80opKSnUtFFFFFFFFFFFFFFFFFFJSUVi6/pYvLY7R81Y/h+5NhMLN+MV2CYIzShRUN3bi4gMbdxXl/ibRXspiVXCVgW0fmShDXrHhi0EFihx2o1rVTB+7h+90rJt9FfU1LXA+lNs7f+ytTS3XhTXZowZMD0rgfHFmyrvxXFQr5ksaV6t4UsRa2g4wSK3sVlazqQtYMRf6yub0zxBc/bMXBOw1u3en297Gs0YG4c1LZD7RatFIOnArz/wAUaa9teb1X5aj0jTP7RwoFdtonhmK0IZ1rpFAVdqDgU7AUUAYFZGu3KxWuO9M0EblD4raoFJ/FSkYrJ1bVobGHqN1eZ61qz30pG7iskbQKQUn0pdpwK2tE0KW+YfL8teh6NoENgo+X5q3R2pmOafjihaWiikyKTilFHHrUbukYy2MVVfVLWP7zCoTr1kP46j/4SOxX+Ok/4SfT+m+nDxFYHo9PXXLE/wAQqdNStW4VhVpGRhxR93hcU2WJJI8OARXHa94XU/v7Rfn9qpaNr0ljOLa54A4rvbaZJ4lZOmKlBA4p1FFLRRRRRRRRRRRRRRRRRRSUU00vGMGuW1fTfJuPtUYxWlpWpI8Sxt94VrfeHBpMNWbrthHeWLAjnHWvNLax8rWPIxXpto6WunKrcYFYSQ/b73PYV1UEQhjUAdBWH4gs2x9qj4K1b8P3Pn2Y3/fqj4xgWSx6V5/otp5uogY+6a9dtUEVsgHpSXUvkWxf0rmIYG1G/wB7cpWvNodqYcJGAwFZentNa3Jilb5OgFdHFAqgFKxvFGnJPZMQvzYrj/DNx9g1MRvwBXpsUoliDrUijAoAqC8uktYDIe1cXLPLq2o7Uz5ddhp1sLW2Vcc1bIzilxSYqhq1+llalj6V5VrOqzXdy2GOysoikUU5h6UmMLWzoWjy30q8fLXp2labHYQKAoBxWkOlA9ulOpKBS0ZoxSYFQzXEduuXNc/qPiq2txhetc/P4zz/AKskVkXPie8l48w4rOfVbp+r1D9sn/vmm/apf71N86T1NPF3MOjU4X1wOj1LHq12nSStGHxNex4/eGtO18YuuPMJzWxZeNbVsLIK3YNatLhMA8VheJNFt54/tNrtDLycVneGNfeK4+zTN7V36lWUMp7U9adRS0UUUUUUUUUUUUUUUUUUlFJilqtdQCeLy65a6024s5vMj6egq/YapIcIyEY9a1vtyhR61lavrYtbVuM1yugj+0db84Diuy1seXZqq8Gq/huPBbIroRVXUYPPtDGO9ZmkKLaXyal8Rx77M/SuI8MQZ1R+Ohr0tBiNB7Vh69eY/wBHWrWhW3l2oYjmtQVzusw+XKrJWppVz50GPQVZuYhLbFcV5Vr0RstULpxiuj8N+IWaNYmU8V1aXquuelVrzXILaPPHFcxPqc+sz+TEGCdK6XRdHSyiBYDdWxjI4oUYpM4qC6u0tbcuxAxXmniXXmvZDFGTt9q5rBHUGmUvalBrQ0fTXv7lUA4r1XRdKjsbZRtG7FagFN2n8KeBS0UlANJUM91Fbrl2ArnNS8XQ2owuG+lcbqPiWa6J2kgGsN5XkPzsTUVFFLSUUUUUuaBT84+7xU0V7PHja54rRTX7gQ+WzEjGKzkmKTeavFeg+E/EHnItvLyfWuwGDyDxTge1LRS0UUUUUUUUUUUUUUUUUlFJRmk25pCqkYK1XazjfooWq/8AZQJ+9XM+Lljt7Yx7vmxUfgaEAK+K6PWDvwtWdLt/KiB9qvrQeeK5h3aLW9o6VtXsXn2ePUVgaDpXkXrtiupk+WD6CuTeNrnVVyDiurhTyoQoFSfw1namg8k8dqreHDuWT2rZPPy15t4zg2XJOKm8HwRTkKcA12TaUNnD4rivEWnn7dHHHKT7V1XhzSktLUF0GfWt3ORxTh0pKa7LGm5u1ef+KtVlubn7Nb5I6cVS03wyTH51wSB71jat5cU5iQA4rMI4oHSrNhbG5nVAOten+G9FS1gVyBmujxgYFIKdiijtSCg0mQF9BWTqWswWUR+cHFec63r017MQhIHtWIzu/wB4n8abikxSUlFFFFFFFFOFJQKKd2q5pd41pOrA16toN+LuxX+9Wv2FOHSiloooooooooooooooooopKTFJijoOKAaPpUFzMsEBdjjAryzxLqP26/8AlOVrtfBtqq6crir1+v7zNaNmwMK/SpwKKwZ7f/iZebWzF/qQKVI0j6YqK8YJCR7VQ06NC+7vWxkUwH5qraiu6HHtVTQECK9aoHNcN44hzGWArA8I3Pk3u0nivTnlA08sD/DXF2azahqgcjKoa7xcCNVHpThwMU8dKTFVr5d9syL3FYNh4fRZfOmHT1rP8U60lrAbWDHHpXn7sZn3MeajpYxk4Fd34S0PO2Z16V3igRoFWn/w0gFOopKWmswRcmuW1/xJHaoY4nFeeX2pS3jck4qkppKM0maSiiiiiiiiilpc0CkFKKOnSuw8G6psuFiduK9JRw6Bh0p9ApaKKKKKKKKKKKKKKKKKSiijFIABS007VHNcb4x1gQp5KN1FcFZR/aLsKeteueHofs+mLHTdY/dxBhT9HlLrWnmk7YrHupQlxitS35gWnFTxVTUlPlcelQaQmErTxzS8LVDUZQiY9aZo6/Ka0cVy3jCH/QjmvOLaY211uWu3/wCEhR7BIVb5sYrV8JWpSN2cdeldKFAoxS4wKAaawrD8Qan9htSOhxXluoXTXU5kJqr2pOTWvoGnm7ukI5Ar1nTrZbS2VQO1WhzzTu1AopKTOKQttXJ6VyPiPxEkCGONueled3Nw9zLukqE8cCjjFNpKKKKKKKKKKKWkpaSloFLVmwnNtOrjivVvDt/9rtEHtW5QtOoooooooooooooopKKKKKKKb0o3Cs3V75LG2bf6cV5Lqt495dMWOQKs+GU36pGK9ft4xHCAB2qprCbrcVW0TrgGtclU68UgcMPlIrBugWvsGtu2G23WpM9MVR1RtsfHpUekZ8vOa09wPTFIcH8KwNVmBmRBWxZw+VEuPSrNc/4uj/4lZryh0PmVe0hGkukU9BXrumQiK1TA7Vb68Ufdpe1HbioJZBbxF3PSvMPFuqm7u9qH5a5sDihRT4o/MlVB3r0zwjpAt4Azryea6n2HSne1LRRSZpNoxXM+KdZW0ttkTYNeY3VzJcylnNQ0mKKKSiiiiiiiiiiiiiilopKcK7HwTqXlXGxzxXo8ZDIGFOFPoooooooooooooopKKKKMUUmOKRePlpk0iwxlj0ArzXxZrP2l/KQ8LXJV0ng633ahG/YV6r0AHtUN3H5sGPaufsHe3vCvOKTxRq5sUXae1Z3hvxH584ic/nWtMwbVMLzXQQj9yBUgAFZ2qj93+Fczca0LOAqpqHQvEzT34iau2MirBv6cVz8K/a7zI7V0yDEYFKOKxPFZH9lkV5JP/rjirGlTeTeRmvYtLuUuLRNh6CrVO4IoFNziuU8W6sIYTEjc4rzN2Lvk80n8NIOK6HwtpjXdwGK8CvVbaNYoFRRjAqQDFLS5oxRTcVnatqKWVueRnFeVa1qLXt03PFZVO6UlFJSUUUUUUUUUUUUUUUUoFLigVasJ2trhSOK9d0G7F1Zp7CtLNSUUUUUUUUUUUUUUUUlJmgGjNApOB1qJ54Y+rgVyXijxAiweVERn2rzuV2lkLU0Iz9BXd+B7YqgYrg132OKO1YOpweT+8UVw3iO/+1OFP8PFYlpcfZLlXTtXeeGme9uFmYHFdqoxS8etUdWQmyfb1AryPU5HF2d3T0q54VTzdXX0r0HWXK2KIh5pPD1s0WWb+Kt3J9KMVh+KB/xLyK8nuOJzUSna2RXdeC9ZK/uXPtXe5BUbaFzTsVBeTCC1YnjivJNfv2ubojOQKx6Snxjc6qK9S8H2QgtgSvOK6YcUo6Ue1HQUo6U3rTJ5Fhi3HjFeX+K9Waa6aJWOK5roKaRSUUUlFFFFFFFFFLRSUtFGKMUo4pKUcClz09q7/wAB6hkeWx9q7rjGafRRRRRRRRRRRRRRSUUyhabJIkK5c4FY914itouEcZrFudeu5R/o67qpp/a12eY2Ap48Ky3DZlBq7D4Lt8ANV+HwfZRDitax0uGyXEdXPp0oxkVFPAs0W1vSuD8ReGX8zMC5rAi8O3Qcb4yBXovhu0jtLBQ2BitWS5jjH3hWVJqy7sKat2863lsyV514q02SG7yq/LVTQJ/sd0DXb2vnX7AkfLXSwRCKFVA6VKKTpVPVLQXVtsri7nwgXfIU1l3PhK6T7kZqpDY32mzq3lkYrt9H8QRmEC5bDCugtr2G5X5CCKnGfwrk/Gmo+RDsQ+xrzRzuYmkHSkHvWlodobq8UDnFevabbLBZxqPSrYpORTqSl7U0Gub8VamLa0ZO+K8ulkM0m89ajGD1pDTaKKKKKKKKKKKKWiiilpwU9hR5belJsI7UgFH0pUA71teGLv7LeKM4zXrkDB4Ub2qaiiiiiiiiiiiiikooqPkn2rM1TVY7VMKfmFczJqF9qL+XHnaat2nhjfhpwa3rXRba3UYWr8aRxjC1JRzRSBgeKTpwKXoKY0yxLl+BWTea5ZR/ebpXPaj4mtMbYiM1mf8ACVOse1CaryeJZGXrUNtrADZet7TNfijkUZ4NdJLb2esQZ43YrPh8JQLLnbxXQ21pHaRBUHSrHpRyKOMUgBP0pcjpijHGMCqk2nW8/wDrFrD1HwzGf+PZaxf+JlpTbedord0/xFD9nxMcHFcN4l1H7ZcEKflFYXbikpfau48D6d++8zHFehKNoxS0tFNFLUMzeVCX9K8v8W6j9pu9o6VzQ4pKM0lFFFFFFFFFFKKKKKcR",
@@ -488,8 +494,6 @@ export default {
         initData.data = dataContent;
         this.initDataArray[el.clientID] = initData;
       };
-
-      //console.log("initDataArray: ", this.initDataArray);
     }
   },
   computed: {
