+++ cccm-frontend/src/components/form.api.ts
@@ -1,3 +1,4 @@
+import { identifier } from '@babel/types';
 import axios from 'axios';
 //import { config } from 'process';
 
@@ -7,10 +8,10 @@ const axiosClient = axios.create({
 });
 
 // function to async fetch the location info
-export async function async_getLocationInfo() {
+export async function async_getUserDefaultLocation() {
     try {
         //console.log("VUE_APP_CCCM_API_ENDPOINT: ", config.VUE_APP_CCCM_API_ENDPOINT);
-        const { data } = await axiosClient.get('/locations');
+        const { data } = await axiosClient.get('/user/location');
         return [null, data];
     }catch (error) {
         return [error];
@@ -18,10 +19,32 @@ export async function async_getLocationInfo() {
 }
 
 // function to sync fetch the location info
-export function getLocationInfo() {
+export function getUserDefaultLocation() {
     try {
         //console.log("VUE_APP_CCCM_API_ENDPOINT: ", config.VUE_APP_CCCM_API_ENDPOINT);
-        const data = axiosClient.get('/locations');
+        const data = axiosClient.get('/user/location');
+        return [null, data];
+    }catch (error) {
+        return [error];
+    }
+}
+
+// function to async fetch the locations info
+export async function async_getUserLocations() {
+    try {
+        //console.log("VUE_APP_CCCM_API_ENDPOINT: ", config.VUE_APP_CCCM_API_ENDPOINT);
+        const { data } = await axiosClient.get('/user/locations');
+        return [null, data];
+    }catch (error) {
+        return [error];
+    }
+}
+
+// function to sync fetch the locations info
+export function getUserLocations() {
+    try {
+        //console.log("VUE_APP_CCCM_API_ENDPOINT: ", config.VUE_APP_CCCM_API_ENDPOINT);
+        const data = axiosClient.get('/user/locations');
         return [null, data];
     }catch (error) {
         return [error];
@@ -70,21 +93,23 @@ export async function createForm(formData: object) {
     }
 }
 
-// function to search client which is used by PO client search and client search
-export async function clientSearch(address: String, age: number, birthYear: number, clientNum: String, 
-    gender: String, location: String, name: String, officer: String, soundex: boolean) {
+// function to search client based on general info 
+export async function clientSearchByGeneralInfo(age: number, birthYear: number, gender: String, 
+    givenName: String, identifier: String, identifierType: String, lastName: String,
+    limitTolocation: boolean, range: number, soundex: boolean) {
     try{
         //console.log("ClientSearch payload");
         const { data } = await axiosClient.get('/clients', {
             params: {
-                address: address,
                 age: age,
                 birthYear: birthYear,
-                clientNum: clientNum,
                 gender: gender,
-                location: location,
-                name: name,
-                officer: officer,
+                givenName: givenName,
+                identifier: identifier,
+                identifierType: identifierType,
+                lastName: lastName,
+                limitTolocation: limitTolocation,
+                range: range,
                 soundex: soundex
             }
         });
@@ -94,8 +119,30 @@ export async function clientSearch(address: String, age: number, birthYear: numb
     }
 }
 
-// Officer search used by supervisor dashboard
-export async function officerSearch(supervisorID: String) {
+// function to search client based on address info 
+export async function clientSearchByAddressInfo(address: String, addressType: String, city: String, 
+    expired: boolean, limitToLocation: boolean, postalCode: String, province: String) {
+    try{
+        //console.log("ClientSearch payload");
+        const { data } = await axiosClient.get('/clients/addressSearch', {
+            params: {
+                address: address,
+                addressType: addressType,
+                city: city,
+                expired: expired,
+                limitToLocation: limitToLocation,
+                postalCode: postalCode,
+                province: province
+            }
+        });
+        return [null, data];
+    } catch (error) {
+        return [error];
+    }
+}
+
+// Supervisor dashboard search
+export async function dashboardSupervisorSearch(supervisorID: String) {
     try{
         //console.log("Officer search by supervisorID: ", supervisorID);
         const { data } = await axiosClient.get(`/dashboard/supervisor/${supervisorID}`);
@@ -105,11 +152,22 @@ export async function officerSearch(supervisorID: String) {
     }
 }
 
+// PO dashboard search
+export async function dashboardPOSearch(poID: String) {
+    try{
+        //console.log("Officer search by supervisorID: ", supervisorID);
+        const { data } = await axiosClient.get(`/dashboard/supervisor/${poID}`);
+        return [null, data];
+    } catch (error) {
+        return [error];
+    }
+}
+
 // function to get client photo
-export async function photoSearch(clientID: String) {
+export async function photoSearch(clientNum: String) {
     try{
         //console.log("Photo search by clientID: ", clientID);
-        const { data } = await axiosClient.get(`/clients/${clientID}/photo`);
+        const { data } = await axiosClient.get(`/clients/${clientNum}/photo`);
         return [null, data];
     } catch (error) {
         return [error];
